Cleaned up styling of new nameserver inputs
Also fixed an issue with extra whitespace at the start
of the page. The <html> tag needs to be on the next line
of the ending ?> of the first block of PHP. Otherwise,
the whole page is shifted down a little bit. Now all the
pages have the same top margin so the h1 header is in the
same place on every page.