Refactored application
Both the enom and namecheap clients implement a common
interface for easier usage. Domains and DNSSEC records
are objects now which the clients return. The clients
DO NOT expose their underlying XML responses anymore.
We've gone up a level of abstraction. Instead they
return either a single literal value like the price
of a domain or an array or Domain or DNSSecKey objects.

All the class files have been moved to a separate folder
with an htaccess file to deny access from client browsers.
Clients should only access the files under the root. Really,
those need to be moved to a subfolder themselves for even
better containment, but we'll save that for later.

The configuration has been modified a bit as well. Each client
has a key in the config with their specific settings. The sample
config has been edited to show the new layout. This make it
so there doesn't need to be a separate server or username
variable for each client.

The stub API has been updated with Namecheap endpoints.